게시글과 댓글의 수정/삭제 권한 부여하기

+++ app/controllers/foods_controller.rb
@@ -1,6 +1,7 @@
 class FoodsController < ApplicationController
-  before_action :login_check
-	skip_before_action :login_check, :only =>  [:posts, :posts_category, :show]
+	before_action :login_check
+	skip_before_action  :login_check, :only => [:posts, :posts_category, :show]
+	
 	def posts
 		@posts = Post.all
 	end
@@ -21,6 +22,7 @@ class FoodsController < ApplicationController
 
   def show#id받아서 Post테이블에서 검색한게 @post에저장됨.
 		@post = Post.find(params[:id])
+		@comment_writer = User.where(id: session[:user_id])[0]
 	end
 
   def write
@@ -28,7 +30,8 @@ class FoodsController < ApplicationController
 
   def write_complete
 		post = Post.new
-	  post.category = params[:post_category]
+		post.user_id = session[:user_id]	
+		post.category = params[:post_category]
 		post.title = params[:post_title]
 		post.content = params[:post_content]
 		if post.save
@@ -42,10 +45,15 @@ class FoodsController < ApplicationController
 
   def edit
   	@post = Post.find(params[:id])
+		if @post.user_id != session[:user_id]
+		flash[:alert] = "수정권한이 없습니다."
+		redirect_to :back
+		end
 	end
 
   def edit_complete
-  	post = Post.new
+  	post = Post.find(params[:id])
+		post.user_id = session[:user_id]
 		post.category = params[:post_category]
 		post.title = params[:post_title]
 		post.content = params[:post_content]
@@ -60,13 +68,19 @@ class FoodsController < ApplicationController
 
   def delete_complete
  		post = Post.find(params[:id])
-		post.destroy
-		flash[:alert] = "삭제되었습니다."
-		redirect_to "/"
- 	end
+		if post.user_id == session[:user_id]
+			post.destroy
+			flash[:alert] = "삭제되었습니다."
+			redirect_to "/"
+ 		else
+			flash[:alert] = "삭제 권한이 없습니다."
+			redirect_to :back
+		end
+	end
 
 	def write_comment_complete
 		comment = Comment.new
+		comment.user_id = session[:user_id]
 		comment.post_id = params[:post_id]
 		comment.content = params[:comment_content]
 		comment.save
@@ -77,8 +91,13 @@ class FoodsController < ApplicationController
 
 	def delete_comment_complete
 		comment = Comment.find(params[:id])
-		comment.destroy
-		flash[:alert] = "댓글이 삭제되었습니다."
-		redirect_to "/foods/show/#{comment.post_id}"
+		if comment.user_id == session[:user_id]
+			comment.destroy
+			flash[:alert] = "댓글이 삭제되었습니다."
+			redirect_to "/foods/show/#{comment.post_id}"
+		else
+			flash[:alert] = "해당 댓글의 삭제 권한이 없습니다."
+			redirect_to :back
+		end
 	end
 end
